fix(logger): fix serviceContext
